refactor(iterator): promote DBIterator state machine to runtime/iterator
Iterator was 539 lines: ~30 lines of DB-coupled constructor + 500
lines of pure state-machine logic (populate/seek/next, range tombstone
filtering, value-pointer materialization). The state machine lives
in runtime/iterator/iterator.go behind a tiny Deps bag (Storage +
Vlog + *runtime.IteratorPool); the root keeps just the (db *DB)
NewIterator / NewInternalIterator facades that build Deps from DB
internals.

stats.go is intentionally NOT promoted: it reads from 13+ distinct
DB fields and has almost no logic of its own beyond per-subsystem
snapshot routing. A stats.Host interface would ~double the surface
area for no real abstraction win — keeping stats.go at root is the
cleaner call here.
